`.setProperty` should not work with camelCase
According to the official spec, `CSSStyleDeclaration.setProperty()`'s
`propertyName` parameter should be the CSS property name in hyphen-case
only. It shouldn't set any styles when the property name is in
camel case or any other form. This is the behavior found in browsers,
so this library should align with it.

https://developer.mozilla.org/en-US/docs/Web/API/CSSStyleDeclaration/setProperty

Closes #60