rgb/rgba more spec compliant, updated hex parsing
Moved / from rgba to simply the alpha case. More spec compliant for a minor change in the regex.
Expanded the colorRegEx1 to capture the hex color string and have case insensitive matching.
Updated the hex parsing to use the regex match group. Logic is mostly the same, I added a check to see if full alpha before adding the alpha value to the list.